refactor(shortcuts): move the whole combo grammar into one neutral module
The combo grammar had two halves in two places: the key tables, the parser
and the matcher in `composables/useKeyboardShortcut.ts`, the display tables
in `components/KeyboardShortcut/combo.ts`. The composable reached into a
component folder for `isMacPlatform`, the only such import in `composables/`.

Both halves now sit in `utils/keyboardShortcutCombo.ts`. `utils` is the layer
both callers already depend on, so the composable no longer imports from a
component folder and no new import runs the other way. The composable keeps
the registry and the lifecycle.

Pure move. No behaviour change and no public API change: `src/index.ts` is
byte-identical. Two internal renames were forced by the merge: the matching
`parseCombo` is now `parseComboForMatching`, because the display half already
owns that name, and the two `warned` dedup sets are now `warnedMatchKeys` and
`warnedDisplayTokens`. The sets stay separate, so each reset helper clears
what it cleared before.

`useKeyboardShortcut.ts` goes from 568 to 320 lines.
