+++ src/composables/useKeyboardShortcut.ts
@@ -6,138 +6,21 @@ import {
   toValue,
   type MaybeRefOrGetter,
 } from 'vue'
-import { isMacPlatform } from '../components/KeyboardShortcut/combo'
+import {
+  isStillHeld,
+  matchesCombo,
+  parseComboForMatching,
+  warnOnce,
+  type KeyboardShortcutCombo,
+} from '../utils/keyboardShortcutCombo'
 
-// ---------------------------------------------------------------------------
-// Key vocabulary
-//
-// A combo is `Mod+Ctrl+Alt+Shift+<Key>`, in that order. Punctuation uses a
-// name, never the character, because `+` is the separator.
-//
-// Letters, function keys and named keys match `event.key`. Digits and
-// punctuation match `event.code`, so `Mod+Shift+Digit1` fires on both the
-// digit and the shifted character the layout puts on that physical key.
-// ---------------------------------------------------------------------------
-
-/** Keys matched against `event.key`. */
-const KEYS_BY_EVENT_KEY = {
-  A: 'a',
-  B: 'b',
-  C: 'c',
-  D: 'd',
-  E: 'e',
-  F: 'f',
-  G: 'g',
-  H: 'h',
-  I: 'i',
-  J: 'j',
-  K: 'k',
-  L: 'l',
-  M: 'm',
-  N: 'n',
-  O: 'o',
-  P: 'p',
-  Q: 'q',
-  R: 'r',
-  S: 's',
-  T: 't',
-  U: 'u',
-  V: 'v',
-  W: 'w',
-  X: 'x',
-  Y: 'y',
-  Z: 'z',
-  F1: 'F1',
-  F2: 'F2',
-  F3: 'F3',
-  F4: 'F4',
-  F5: 'F5',
-  F6: 'F6',
-  F7: 'F7',
-  F8: 'F8',
-  F9: 'F9',
-  F10: 'F10',
-  F11: 'F11',
-  F12: 'F12',
-  Escape: 'Escape',
-  Enter: 'Enter',
-  Tab: 'Tab',
-  Space: ' ',
-  Backspace: 'Backspace',
-  Delete: 'Delete',
-  Insert: 'Insert',
-  ArrowUp: 'ArrowUp',
-  ArrowDown: 'ArrowDown',
-  ArrowLeft: 'ArrowLeft',
-  ArrowRight: 'ArrowRight',
-  Home: 'Home',
-  End: 'End',
-  PageUp: 'PageUp',
-  PageDown: 'PageDown',
-} as const
-
-/** Keys matched against `event.code`, so a shifted character still resolves. */
-const KEYS_BY_EVENT_CODE = {
-  Digit0: 'Digit0',
-  Digit1: 'Digit1',
-  Digit2: 'Digit2',
-  Digit3: 'Digit3',
-  Digit4: 'Digit4',
-  Digit5: 'Digit5',
-  Digit6: 'Digit6',
-  Digit7: 'Digit7',
-  Digit8: 'Digit8',
-  Digit9: 'Digit9',
-  /** The keypad `+`. The `+` a layout types with Shift is `Shift+Equal`. */
-  Plus: 'NumpadAdd',
-  Minus: 'Minus',
-  Equal: 'Equal',
-  Slash: 'Slash',
-  Backslash: 'Backslash',
-  Backtick: 'Backquote',
-  Comma: 'Comma',
-  Period: 'Period',
-  Semicolon: 'Semicolon',
-  Quote: 'Quote',
-  BracketLeft: 'BracketLeft',
-  BracketRight: 'BracketRight',
-} as const
-
-// Which table a key sits in is an internal detail: an app names a key with
-// `KeyboardShortcutKey`, or a whole combo with `KeyboardShortcutCombo`. Both
-// halves stay module-internal so the split is free to move after 1.0.0 (P13).
-
-/** A key name matched against `event.key`. */
-type KeyboardShortcutNamedKey = keyof typeof KEYS_BY_EVENT_KEY
-/** A key name matched against `event.code`. */
-type KeyboardShortcutCodedKey = keyof typeof KEYS_BY_EVENT_CODE
-/** Every key name a combo accepts. */
-export type KeyboardShortcutKey =
-  | KeyboardShortcutNamedKey
-  | KeyboardShortcutCodedKey
-
-/**
- * @internal Exported for unit tests only: every key name a combo accepts, so a
- * test can walk them through the display grammar in
- * `components/KeyboardShortcut/combo.ts` and catch the two tables drifting.
- */
-export const _keyboardShortcutKeyNames = [
-  ...Object.keys(KEYS_BY_EVENT_KEY),
-  ...Object.keys(KEYS_BY_EVENT_CODE),
-] as KeyboardShortcutKey[]
-
-type ModPrefix = '' | 'Mod+'
-type CtrlPrefix = '' | 'Ctrl+'
-type AltPrefix = '' | 'Alt+'
-type ShiftPrefix = '' | 'Shift+'
-
-/**
- * A key combination, written `Mod+Ctrl+Alt+Shift+<Key>` with the modifiers in
- * that order. `Mod` is Cmd on macOS and Ctrl elsewhere; `Ctrl` is Control on
- * every platform.
- */
-export type KeyboardShortcutCombo =
-  `${ModPrefix}${CtrlPrefix}${AltPrefix}${ShiftPrefix}${KeyboardShortcutKey}`
+// The combo grammar — the key tables, the parser, the matcher — lives in
+// `utils/keyboardShortcutCombo.ts`, next to the display half the chip draws
+// with. This file is the registry and the lifecycle around it.
+export type {
+  KeyboardShortcutCombo,
+  KeyboardShortcutKey,
+} from '../utils/keyboardShortcutCombo'
 
 // ---------------------------------------------------------------------------
 // Config
@@ -187,137 +70,6 @@ export interface HoldKeyboardShortcutConfig extends KeyboardShortcutBase {
 export type KeyboardShortcutConfig =
   | PressKeyboardShortcutConfig
   | HoldKeyboardShortcutConfig
-
-// ---------------------------------------------------------------------------
-// Combo parsing
-// ---------------------------------------------------------------------------
-
-interface ParsedCombo {
-  ctrl: boolean
-  meta: boolean
-  alt: boolean
-  shift: boolean
-  key: KeyboardShortcutKey
-}
-
-const KEY_BY_LOWERCASE_NAME = new Map<string, KeyboardShortcutKey>(
-  [...Object.keys(KEYS_BY_EVENT_KEY), ...Object.keys(KEYS_BY_EVENT_CODE)].map(
-    (name) => [name.toLowerCase(), name as KeyboardShortcutKey],
-  ),
-)
-
-const parseCache = new Map<string, ParsedCombo | null>()
-
-/** @internal Exported for unit tests only. */
-export function parseCombo(combo: string): ParsedCombo | null {
-  // A JavaScript call site can hand us anything, including the v0 config that
-  // had no `combo` at all. Warn rather than throw from a global listener.
-  if (typeof combo !== 'string') {
-    warnOnce(
-      'combo:missing',
-      `A shortcut was registered without a combo, so it will never fire. ` +
-        `Replace the v0 key/ctrl/shift/alt fields with one combo string, ` +
-        `e.g. { key: 's', ctrl: true } becomes { combo: 'Mod+S' }.`,
-    )
-    return null
-  }
-
-  const cached = parseCache.get(combo)
-  if (cached !== undefined) return cached
-
-  const parts = combo.split('+').map((part) => part.trim())
-  const rawKey = parts.pop() ?? ''
-  const key = KEY_BY_LOWERCASE_NAME.get(rawKey.toLowerCase())
-
-  let parsed: ParsedCombo | null = key
-    ? { ctrl: false, meta: false, alt: false, shift: false, key }
-    : null
-
-  if (parsed) {
-    for (const modifier of parts) {
-      switch (modifier.toLowerCase()) {
-        case 'mod':
-          // The chip reads the same check, so what a row draws is what fires.
-          if (isMacPlatform()) parsed.meta = true
-          else parsed.ctrl = true
-          break
-        case 'ctrl':
-          parsed.ctrl = true
-          break
-        case 'alt':
-          parsed.alt = true
-          break
-        case 'shift':
-          parsed.shift = true
-          break
-        default:
-          parsed = null
-      }
-      if (!parsed) break
-    }
-  }
-
-  if (!parsed) {
-    warnOnce(
-      `combo:${combo}`,
-      `"${combo}" is not a valid shortcut combo, so it will never fire. ` +
-        `Write the modifiers as Mod+Ctrl+Alt+Shift and name the key ` +
-        `(e.g. "Mod+Shift+K", "Mod+Slash", "Mod+Shift+Digit1").`,
-    )
-  }
-
-  parseCache.set(combo, parsed)
-  return parsed
-}
-
-function matchesKey(e: KeyboardEvent, key: KeyboardShortcutKey): boolean {
-  if (key in KEYS_BY_EVENT_CODE) {
-    return e.code === KEYS_BY_EVENT_CODE[key as KeyboardShortcutCodedKey]
-  }
-  const expected = KEYS_BY_EVENT_KEY[key as KeyboardShortcutNamedKey]
-  return e.key.toLowerCase() === expected.toLowerCase()
-}
-
-/** @internal Exported for unit tests only. */
-export function matchesCombo(e: KeyboardEvent, combo: string): boolean {
-  const parsed = parseCombo(combo)
-  if (!parsed) return false
-  if (e.ctrlKey !== parsed.ctrl) return false
-  if (e.metaKey !== parsed.meta) return false
-  if (e.altKey !== parsed.alt) return false
-  if (e.shiftKey !== parsed.shift) return false
-  return matchesKey(e, parsed.key)
-}
-
-// A held shortcut ends when its own key goes up, or when a modifier it needs
-// goes up. Modifiers it does not need are ignored, so releasing Shift on a
-// `Mod+L` hold keeps it running.
-function isStillHeld(e: KeyboardEvent, parsed: ParsedCombo): boolean {
-  if (matchesKey(e, parsed.key)) return false
-  if (parsed.ctrl && !e.ctrlKey) return false
-  if (parsed.meta && !e.metaKey) return false
-  if (parsed.alt && !e.altKey) return false
-  if (parsed.shift && !e.shiftKey) return false
-  return true
-}
-
-// ---------------------------------------------------------------------------
-// Dev warnings, once per key per page load
-// ---------------------------------------------------------------------------
-
-const warned = new Set<string>()
-
-function warnOnce(key: string, message: string) {
-  if (import.meta.env.PROD) return
-  if (warned.has(key)) return
-  warned.add(key)
-  console.warn(`[frappe-ui] ${message}`)
-}
-
-/** Test-only: clear the dedup set so each test sees a fresh warning surface. */
-export function _resetKeyboardShortcutWarnings() {
-  warned.clear()
-}
 
 // ---------------------------------------------------------------------------
 // Registry (module-internal)
@@ -391,7 +143,7 @@ function globalKeydownHandler(e: KeyboardEvent) {
 function globalKeyupHandler(e: KeyboardEvent) {
   for (const { id, config } of [...registrations]) {
     if (!heldShortcuts.has(id)) continue
-    const parsed = parseCombo(config.combo)
+    const parsed = parseComboForMatching(config.combo)
     if (!parsed || isStillHeld(e, parsed)) continue
     heldShortcuts.delete(id)
     config.onRelease?.(e)
